+++ packages/client/.gitignore
@@ -11,7 +11,6 @@
 # production
 /build
 /dist
-/public/devtools.js
 
 # misc
 .DS_Store
