+++ packages/client/package.json
@@ -5,29 +5,13 @@
 	"private": true,
 	"homepage": "https://podium-lib.io/",
 	"scripts": {
-		"build": "npm run build:js && npm run build:web-ext",
-		"build:js": "rollup --config rollup.config.js",
-		"build:web-ext": "web-ext build --source-dir public --artifacts-dir dist --overwrite-dest",
-		"start": "npm-run-all --parallel watch:js watch:web-ext",
-		"start:chrome": "npm-run-all --parallel watch:js watch:web-ext-chrome",
-		"watch:js": "rollup --config rollup.config.js --watch --watch.onEnd=\"web\"",
-		"watch:web-ext": "web-ext run --source-dir public --devtools",
-		"watch:web-ext-chrome": "web-ext run --target chromium --source-dir public --devtools"
-	},
-	"dependencies": {
-		"classnames": "2.5.1",
-		"compare-versions": "6.1.0",
-		"react": "18.2.0",
-		"react-dom": "18.2.0"
+		"build": "web-ext build --source-dir public --artifacts-dir dist --overwrite-dest",
+		"start": "npm run web-ext-firefox",
+		"start:chrome": "npm run web-ext-chrome",
+		"web-ext-firefox": "web-ext run --source-dir public --devtools",
+		"web-ext-chrome": "web-ext run --target chromium --source-dir public --devtools"
 	},
 	"devDependencies": {
-		"@babel/preset-react": "7.24.1",
-		"@rollup/plugin-babel": "6.0.4",
-		"@rollup/plugin-commonjs": "25.0.7",
-		"@rollup/plugin-node-resolve": "15.2.3",
-		"@rollup/plugin-replace": "5.0.5",
-		"npm-run-all": "4.1.5",
-		"rollup": "4.13.1",
 		"web-ext": "7.11.0"
 	}
 }
