+++ packages/client/public/extension.js
@@ -1,2 +1,2 @@
-const useBroswer = chrome || browser;
-useBroswer.devtools.panels.create("Podium", "icon.png", "index.html");
+const useBrowser = chrome || browser;
+useBrowser.devtools.panels.create("Podium", "icon.png", "index.html");
