+++ packages/client/public/index.html
@@ -3,16 +3,23 @@
 	<head>
 		<meta charset="utf-8" />
 		<title>Podium Developer Tools</title>
-		<link
-			rel="stylesheet"
-			href="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/css/bootstrap.min.css"
-			integrity="sha384-MCw98/SFnGE8fJT3GXwEOngsV7Zt27NXFoaoApmYm81iuXoPkFOJwJ8ERdknLPMO"
-			crossorigin="anonymous"
-		/>
 		<link rel="stylesheet" href="./devtools.css" />
 	</head>
 	<body>
-		<div id="root"></div>
+		<main class="podium">
+			<h1>Podium Developer Tools</h1>
+			<div class="status">
+				<p id="status-message"></p>
+				<div id="status-meta"></div>
+			</div>
+			<form id="connection">
+				<label>
+					DevTools URL
+					<input type="text" name="apibase" value="http://localhost:8172/" />
+				</label>
+				<button type="submit">Connect</button>
+			</form>
+		</main>
 		<script src="./devtools.js" type="module"></script>
 	</body>
 </html>
