--- packages/client/src/index.jsx
@@ -1,10 +1,0 @@
-import React from "react";
-import ReactDOM from "react-dom/client";
-import App from "./components/app/App.jsx";
-
-const root = ReactDOM.createRoot(document.getElementById("root"));
-root.render(
-	<React.StrictMode>
-		<App />
-	</React.StrictMode>,
-);
